Refactor WhiteboardCanvas to support custom DOM.
I've moved most logic related to modifying the fabric canvas away from the WhiteboardContext. Instead it's been moved to the WhiteboardCanvas component. This allow us to have the Toolbar and Canvas in separate locations in the page DOM. I had to create an interface for "registering" the different actions in the context.