Add concept of generatedBy.
Add concept of generatedBy to be able to better determine if events emitted from the canvas should be serialized and sent over the network. There's already an ID field for each object but the ID field will remain the same each canvas reload. Which means objects created in earlier session either wont show up because the events is filtered based on ID field, or if they're removed for example, the original user wont see that they're removed after reloading.